Separação
Separação das funções do core.py e das funções que fazem as requisições, isso deixa o código mais organizado, facilita a manutenção e abre caminho para a escrita de testes unitários.
